Add ImperativeDialog - fix bug while keeping simple
